refactor: change Category id type from Long to Integer

+++ src/main/java/com/devsuperior/sistemaevento/entities/Category.java
@@ -11,7 +11,7 @@ public class Category implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(columnDefinition = "TEXT")
     private String description;
@@ -19,16 +19,16 @@ public class Category implements Serializable {
     public Category() {
     }
 
-    public Category(Long id, String description) {
+    public Category(Integer id, String description) {
         this.id = id;
         this.description = description;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
